// src/types/client-supplier/client-supplier.ts
export type ClientSupplierRecord = Client | Supplier

export interface ClientSupplierFormValues {
  name: string
  cuit: string
}

export interface ClientSupplierModalInitialValues {
  name?: string
  cuit?: string
}